Unify RAG script and add seeding option
